pyproject.toml
[project]
name = "polymarket-watcher"
version = "0.1.0"
description = "Monitor Polymarket for anomalous trading behavior"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    "websockets>=15.0.1",
]

[project.scripts]
polymarket-watcher = "src.main:main"

[dependency-groups]
dev = []
